refactor(citations): extract format-agnostic Citation model into citation.py
The Citation value object, Harvard renderers, and shared helpers lived in
csl.py, so bibtex.py imported the shared model and a private clean_doi from a
file named after the other input format. Move them into citation.py; csl.py and
bibtex.py are now symmetric adapters (format in, Citation out) that both import
the public Citation and clean_doi. Pure move/rename, no behavior change.
